+++ src/Service/CognitoIdentityProvider/tests/Unit/Input/AdminDisableUserRequestTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace AsyncAws\CognitoIdentityProvider\Tests\Unit\Input;
+
+use AsyncAws\CognitoIdentityProvider\Input\AdminDisableUserRequest;
+use AsyncAws\Core\Test\TestCase;
+
+class AdminDisableUserRequestTest extends TestCase
+{
+    public function testRequest(): void
+    {
+        $input = new AdminDisableUserRequest([
+            'UserPoolId' => 'us-east-1_1337oL33t',
+            'Username' => '1c202820-8eb5-11ea-bc55-0242ac130003',
+        ]);
+
+        // see https://docs.aws.amazon.com/cognito-user-identity-pools/latest/APIReference/API_AdminDisableUser.html
+        $expected = '
+                POST / HTTP/1.0
+                Content-Type: application/x-amz-json-1.1
+                X-AMZ-Target: AWSCognitoIdentityProviderService.AdminDisableUser
+
+                {
+                    "UserPoolId": "us-east-1_1337oL33t",
+                    "Username": "1c202820-8eb5-11ea-bc55-0242ac130003"
+                }
+                ';
+
+        self::assertRequestEqualsHttpRequest($expected, $input->request());
+    }
+}