+++ src/Service/CognitoIdentityProvider/tests/Unit/Result/AdminEnableUserResponseTest.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace AsyncAws\CognitoIdentityProvider\Tests\Unit\Result;
+
+use AsyncAws\CognitoIdentityProvider\Result\AdminEnableUserResponse;
+use AsyncAws\Core\Response;
+use AsyncAws\Core\Test\Http\SimpleMockedResponse;
+use AsyncAws\Core\Test\TestCase;
+use Psr\Log\NullLogger;
+use Symfony\Component\HttpClient\MockHttpClient;
+
+class AdminEnableUserResponseTest extends TestCase
+{
+    public function testAdminEnableUserResponse(): void
+    {
+        // see https://docs.aws.amazon.com/cognito-user-identity-pools/latest/APIReference/API_AdminEnableUser.html
+        $response = new SimpleMockedResponse('{}');
+
+        $client = new MockHttpClient($response);
+        $result = new AdminEnableUserResponse(new Response($client->request('POST', 'http://localhost'), $client, new NullLogger()));
+        self::assertTrue($result->resolve());
+    }
+}